+++ src/index.html
@@ -19,6 +19,9 @@
         <a href="#">About</a>
       </div>
     </header>
+    <form>
+      <input type="text" placeholder="Enter a category..." class="input-category">
+    </form>
     <div class="show-container">
     </div>
     <footer>
